Revert WebSocketProvider and WebSocketAwarenessProvider move
